requirements.txt
aiodns==3.2.0
aiofiles==23.2.1
aiohttp==3.9.5
aiomultiprocess==0.9.1
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.6.0
anyio==4.3.0
apt-clone==0.2.1
apturl==0.5.2
argcomplete==3.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.1
attrs==23.2.0
babel==2.16.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
censys==2.2.12
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
click-plugins==1.1.1
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
configobj==5.0.6
contourpy==1.2.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.7
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.7.0
dj-rest-auth==7.0.0
Django==5.1.3
django-allauth==65.2.0
django-colorfield==0.11.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
dnspython==2.6.1
drf-yasg==1.21.8
email_validator==2.1.1
exceptiongroup==1.2.1
executing==2.1.0
eyeD3==0.8.10
fastapi==0.111.0
fastapi-cli==0.0.3
fastjsonschema==2.20.0
filelock==3.6.0
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.4.1
greenlet==3.0.3
grpcio==1.30.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
idna==3.3
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
importlib_resources==6.4.0
inflection==0.5.1
ipykernel==6.29.5
ipython==8.28.0
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
limits==3.12.0
louis==3.20.0
lxml==5.2.1
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
more-itertools==8.10.0
multidict==6.0.5
mutagen==1.45.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nemo-emblems==6.0.1
nest-asyncio==1.6.0
netaddr==1.2.1
netifaces==0.11.0
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.0
onboard==1.4.1
orjson==3.10.3
overrides==7.7.0
packaging==24.1
PAM==0.4.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.3.6
playwright==1.44.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycairo==1.20.1
pycares==4.4.0
pycparser==2.22
pycryptodomex==3.11.0
pycups==2.0.1
pycurl==7.44.1
pydantic==2.7.1
pydantic_core==2.18.2
pyee==11.1.0
pyelftools==0.27
Pygments==2.18.0
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-gnupg==0.4.8
python-json-logger==2.0.7
python-magic==0.4.24
python-multipart==0.0.9
python-slugify==8.0.4
python-xlib==0.29
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==6.0.1
pyzmq==26.2.0
qrcode==7.3.1
referencing==0.35.1
reportlab==3.6.8
requests==2.31.0
requests-file==1.5.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.20.0
SecretStorage==3.3.1
Send2Trash==1.8.3
setproctitle==1.2.2
shellingham==1.5.4
shodan==1.31.0
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.3.1
sqlparse==0.5.2
stack-data==0.6.3
starlette==0.37.2
systemd-python==234
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.1.1
tldextract==3.1.2
tomli==2.0.2
tornado==6.4.1
traitlets==5.14.3
typer==0.12.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.11.0
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.10.0
Unidecode==1.3.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.5
uvicorn==0.29.0
uvloop==0.19.0
vboxapi==1.0
wadllib==1.3.6
watchfiles==0.21.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
WoeUSB-ng @ file:///home/de-elite/WoeUSB-ng
wrapt==1.16.0
wxPython==4.0.7
xdg==5
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==3.2.0
yarl==1.9.4
yt-dlp==2022.4.8
zipp==1.0.0
